requirements.txt
django == 5.0.6
python-dotenv == 1.0.1
social-auth-core == 4.5.3
social-auth-app-django == 5.4.0
Pillow == 10.3.0
google.auth == 2.28.1
google-api-python-client == 2.120.0
google-auth-oauthlib == 1.1.0
django-maintenance-mode == 0.21.1
pandas == 2.2.1
djangorestframework == 3.15.1